refactor(html): unify RendererContext + Component::Base via RendererDelegation
RendererContext (used by Drop factories) and Component::Base (used by
Component renderers) both narrowed BaseRenderer surface via forwarders.
Six forwarders were duplicated verbatim across the two classes:
safe_attr, escape_html, extract_block_label, extract_plain_text,
render_mixed_inline, render_liquid.

Extract RendererDelegation module. Both classes include it. The shared
forwarding surface is declared in one place. Per-class forwarders
(RendererContext has 17 render_* methods; Component::Base has 3
specific helpers) stay declared in each class.

Adding a new shared forwarder is one edit in RendererDelegation, not
two coordinated edits across two files.

The 84 one-liner delegations in BaseRenderer itself stay - those
maintain BaseRenderer as the unified entry surface for callers and
earn their keep. The duplication this commit addresses was between
the two narrowing wrappers, not BaseRenderer itself.

Behavior preserved: OIML R 60-1 HTML byte-identical to pre-change
output (only footer timestamp differs). 718 specs pass.

Candidate 6 of the round-2 architecture review.
